Split msm-cp into msm-controller and msm-network-controller
