+++ internal/core/wire_gen.go
@@ -8,8 +8,6 @@ package core
 import (
 	"github.com/media-streaming-mesh/msm-cp/internal/config"
 	node_mapper "github.com/media-streaming-mesh/msm-cp/pkg/node-mapper"
-	stream_mapper "github.com/media-streaming-mesh/msm-cp/pkg/stream-mapper"
-	"sync"
 )
 
 // Injectors from wire.go:
@@ -19,20 +17,11 @@ func InitApp() (*App, error) {
 	grpcImpl := New(cfg)
 
 	nodeMapper := node_mapper.InitializeNodeMapper(cfg)
-	go func() {
-		nodeMapper.WatchNode()
-	}()
-
-	streamMapper := stream_mapper.NewStreamMapper(cfg.Logger, new(sync.Map))
-	go func() {
-		streamMapper.WatchStream()
-	}()
 
 	app := &App{
 		cfg:     cfg,
 		grpcImpl: grpcImpl,
 		nodeMapper: nodeMapper,
-		streamMapper: streamMapper,
 	}
 	return app, nil
 }
